--- ai_service/src/main/java/ai_server/AiServiceApplication.java
@@ -1,17 +1,0 @@
-package ai_server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableFeignClients
-@EnableJpaAuditing
-public class AiServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AiServiceApplication.class, args);
-    }
-
-}
