--- ai_service/src/main/java/ai_server/application/dto/response/AiDeadlineResponseV1.java
@@ -1,16 +1,0 @@
-package ai_server.application.dto.response;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class AiDeadlineResponseV1 {
-
-    private UUID aiLogId;
-    private String orderInfo;
-    private LocalDateTime finalDeadline;
-    private String slackFormattedText;
-}
