--- ai_service/src/main/java/ai_server/common/error/BusinessException.java
@@ -1,14 +1,0 @@
-package ai_server.common.error;
-
-import lombok.Getter;
-
-@Getter
-public class BusinessException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
