--- ai_service/src/main/java/ai_server/common/error/ErrorResponse.java
@@ -1,8 +1,0 @@
-package ai_server.common.error;
-
-import lombok.Builder;
-
-@Builder
-public record ErrorResponse(int status, String code, String message, String path) {
-
-}
