--- ai_service/src/main/java/ai_server/common/success/BaseStatus.java
@@ -1,6 +1,0 @@
-package ai_server.common.success;
-
-public enum BaseStatus {
-
-    OK, CREATED
-}
