--- ai_service/src/main/java/ai_server/infra/config/OpenAiConfig.java
@@ -1,16 +1,0 @@
-package ai_server.infra.config;
-
-import org.springframework.ai.chat.client.ChatClient;
-import org.springframework.ai.openai.OpenAiChatModel;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenAiConfig {
-
-    @Bean
-    public ChatClient chatClient(OpenAiChatModel chatModel) {
-
-        return ChatClient.builder(chatModel).build();
-    }
-}
