--- ai_service/src/main/java/ai_server/infra/config/OpenAiConstants.java
@@ -1,67 +1,0 @@
-package ai_server.infra.config;
-
-public class OpenAiConstants {
-
-    private OpenAiConstants() {}
-
-    public static final String PROVIDER_OPENAI = "OpenAI";
-    public static final String MODEL_GPT_4O_MINI = "gpt-4o-mini";
-
-    /**
-     * 배송 발송 시한 계산 + Slack 메시지 생성용 프롬프트
-     */
-    public static final String DEADLINE_SLACK_PROMPT =
-        """
-        당신은 물류 배송 스케줄러입니다.
-        아래 주문 정보를 바탕으로 **Slack 메시지 형식**으로 발송 안내문을 생성하세요.
-    
-        ✅ 출력 규칙
-        1. 예시 형식을 정확히 따라야 합니다 (줄바꿈 포함).
-        2. 불필요한 설명/서론 없이 **바로 메시지 본문만 출력**합니다.
-        3. 마지막 줄은 다음 템플릿을 반드시 포함해야 합니다:
-           "위 내용을 기반으로 도출된 최종 발송 시한은 yyyy-MM-dd HH:mm 입니다."
-        4. 날짜/시간은 한국 기준(Asia/Seoul)으로 계산한다고 가정합니다.
-        5. 발송 시한 계산은 **도착지 요청 시간 → 경유지 → 발송지** 이동 시간을 고려해 합리적으로 추정합니다.
-    
-        ✅ 출력 예시
-    
-        예시 입력:
-        - 주문 번호: 1
-        - 주문자: 김말숙 / msk@seafood.world
-        - 주문 시각: 2025-12-08 10:00:00
-        - 상품: 마른 오징어 50박스 / 수량 50개
-        - 요청 사항: 12월 12일 3시까지는 보내주세요!
-        - 발송지: 경기 북부 센터
-        - 경유지: 대전광역시 센터, 부산광역시 센터
-        - 도착지: 부산시 사하구 낙동대로 1번길 1 해산물월드
-        - 배송 담당자: 고길동 / kdk@sparta.world
-    
-        예시 출력:
-        주문 번호: 1
-        주문자 정보: 김말숙 / msk@seafood.world
-        주문 시각: 2025-12-08 10:00:00
-        상품 정보: 마른 오징어 50박스
-        요청 사항: 12월 12일 3시까지는 보내주세요!
-        발송지: 경기 북부 센터
-        경유지: 대전광역시 센터, 부산광역시 센터
-        도착지: 부산시 사하구 낙동대로 1번길 1 해산물월드
-        배송담당자: 고길동 / kdk@sparta.world
-    
-        위 내용을 기반으로 도출된 최종 발송 시한은 2025-12-10 09:00 입니다.
-    
-    
-        ✅ 이제 다음 주문 정보를 기반으로 동일한 형식으로 출력하세요.
-    
-        - 주문 번호: %d
-        - 주문자: %s / %s
-        - 주문 시각: %s
-        - 상품: %s / 수량 %d개
-        - 요청 사항: %s
-        - 발송지: %s
-        - 경유지: %s
-        - 도착지: %s
-        - 배송 담당자: %s / %s
-        """;
-}
-
-// TODO - Few-shot Prompting 으로 수정 필요
