+++ ai_service/src/main/java/ai_service/AiServiceApplication.java
@@ -1,0 +1,17 @@
+package ai_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableFeignClients
+@EnableJpaAuditing
+public class AiServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AiServiceApplication.class, args);
+    }
+
+}
