+++ ai_service/src/main/java/ai_service/application/dto/response/AiDeadlineResponseV1.java
@@ -1,0 +1,16 @@
+package ai_service.application.dto.response;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class AiDeadlineResponseV1 {
+
+    private UUID aiLogId;
+    private String orderInfo;
+    private LocalDateTime finalDeadline;
+    private String slackFormattedText;
+}
