+++ ai_service/src/main/java/ai_service/common/error/BusinessException.java
@@ -1,0 +1,14 @@
+package ai_service.common.error;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
