+++ ai_service/src/main/java/ai_service/common/error/ErrorResponse.java
@@ -1,0 +1,8 @@
+package ai_service.common.error;
+
+import lombok.Builder;
+
+@Builder
+public record ErrorResponse(int status, String code, String message, String path) {
+
+}
