+++ ai_service/src/main/java/ai_service/common/success/BaseStatus.java
@@ -1,0 +1,6 @@
+package ai_service.common.success;
+
+public enum BaseStatus {
+
+    OK, CREATED
+}
