+++ ai_service/src/main/java/ai_service/infra/config/OpenAiConfig.java
@@ -1,0 +1,16 @@
+package ai_service.infra.config;
+
+import org.springframework.ai.chat.client.ChatClient;
+import org.springframework.ai.openai.OpenAiChatModel;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenAiConfig {
+
+    @Bean
+    public ChatClient chatClient(OpenAiChatModel chatModel) {
+
+        return ChatClient.builder(chatModel).build();
+    }
+}
