+++ gateway/src/main/resources/application.yml
@@ -19,7 +19,7 @@ spring:
                 - Path=/v1/user/**,/v1/master/**
 
             - id: ai-service
-              uri: lb://ai-server
+              uri: lb://ai-service
               predicates:
                 - Path=/v1/ai/**
